+++ tests/test_main.py
@@ -1,6 +1,7 @@
 import os
 import subprocess
 import sys
+import textwrap
 
 import pytest
 
@@ -673,7 +674,7 @@ def test_newline_preserved(runner, tmp_path, file, newline):
 
 
 def test_no_format_python_code_blocks(runner):
-    file = ".. code-block:: python\n\n" "    def example_function():\n"
+    file = ".. code-block:: python\n\n    def example_function():\n"
     args = ["-t", "rst", "-l", 80, "-r", file, "--no-format-python-code-blocks"]
     result = runner.invoke(main, args=args)
     assert result.exit_code == 0
@@ -700,6 +701,300 @@ def test_comment_preserve_single_line(runner):
     file = "..  A comment in a single line is not placed on the next one.\n"
     fixed = ".. A comment in a single line is not placed on the next one.\n"
     args = ["-r", file]
+    result = runner.invoke(main, args=args)
+    assert result.exit_code == 0
+    assert result.output == fixed
+
+
+def test_section_invalid_adornments(runner):
+    file = """
+              ===
+              One
+              ===
+
+              Two
+              ---
+              Some content.
+           """
+
+    file = textwrap.dedent(file).lstrip()
+    args = ["-s", "#*|=-^\"'~+.`_:# ", "-r", file]
+    result = runner.invoke(main, args=args)
+    assert result.exit_code == 2
+    assert "Section adornments must be unique" in result.output
+
+
+def test_section_reformatting(runner):
+    file = """
+              ===
+              One
+              ===
+
+              Two
+              ---
+
+              Three
+              ~~~~~
+
+              Four
+              ++++
+
+              Five
+              ....
+
+              Six
+              '''
+
+              Two again
+              ---------
+
+              Some content.
+           """
+
+    fixed = """
+               =====
+                One
+               =====
+
+               Two
+               ---
+
+               Three
+               ~~~~~
+
+               Four
+               ++++
+
+               Five
+               ....
+
+               Six
+               '''
+
+               Two again
+               ---------
+
+               Some content.
+            """
+
+    file = textwrap.dedent(file).lstrip()
+    fixed = textwrap.dedent(fixed).lstrip()
+    args = ["-r", file]
+    result = runner.invoke(main, args=args)
+    assert result.exit_code == 0
+    assert result.output == fixed
+
+
+def test_section_reformatting_python_adornments(runner):
+    file = '''
+    """This is an example python file."""
+
+
+    class ExampleClass:
+        """This is a class docstring example.
+
+        ===
+        One
+        ===
+
+        Two
+        ---
+
+        Three
+        ~~~~~
+
+        Four
+        ++++
+
+        Five
+        ....
+
+        Six
+        \'''
+
+        Two again
+        ---------
+
+        Some content.
+
+        """
+
+    '''
+
+    fixed = '''
+    """This is an example python file."""
+
+
+    class ExampleClass:
+        """This is a class docstring example.
+
+        #####
+         One
+        #####
+
+        *****
+         Two
+        *****
+
+        Three
+        =====
+
+        Four
+        ----
+
+        Five
+        ^^^^
+
+        Six
+        """
+
+        ***********
+         Two again
+        ***********
+
+        Some content.
+
+        """
+
+    '''
+
+    file = textwrap.dedent(file).lstrip()
+    fixed = textwrap.dedent(fixed).lstrip()
+    args = ["-t", "py", "-or", file]
+    result = runner.invoke(main, args=args)
+    assert result.exit_code == 0
+    assert result.output == fixed
+
+
+def test_section_reformatting_custom_adornments(runner):
+    file = """
+              ===
+              One
+              ===
+
+              Two
+              ---
+
+              Three
+              ~~~~~
+
+              Four
+              ++++
+
+              Five
+              ....
+
+              Six
+              '''
+
+              Two again
+              ---------
+
+              Some content.
+           """
+
+    fixed = """
+               One
+               ###
+
+               Two
+               ***
+
+               Three
+               =====
+
+               Four
+               ----
+
+               Five
+               ^^^^
+
+               Six
+               \"\"\"
+
+               Two again
+               *********
+
+               Some content.
+            """
+
+    file = textwrap.dedent(file).lstrip()
+    fixed = textwrap.dedent(fixed).lstrip()
+    args = ["-s", '#*=-^"', "-r", file]
+    result = runner.invoke(main, args=args)
+    assert result.exit_code == 0
+    assert result.output == fixed
+
+
+def test_section_reformatting_insufficient_adornments(runner):
+    file = """
+              ===
+              One
+              ===
+
+              Two
+              ---
+
+              Three
+              ~~~~~
+              Some content.
+           """
+
+    file = textwrap.dedent(file).lstrip()
+    args = ["-s", "=:", "-r", file, "-o"]
+    result = runner.invoke(main, args=args)
+    assert result.exit_code == 1
+    assert "there are only 2 adornments to pick from" in result.output
+
+
+def test_section_reformatting_numpydoc(runner):
+    file = """
+              def function(param1: str, param2: int) -> None:
+                  \"\"\"This is a docstring for a python function.
+
+                  This is the summary.
+
+                  Parameters
+                  ----------
+                  param1
+                     Description for param1.
+                  param2
+                     Description for param2.
+
+                  Returns
+                  -------
+                     Absolutely nothing.
+                  \"\"\"
+                  pass
+           """
+
+    fixed = """
+               def function(param1: str, param2: int) -> None:
+                   \"\"\"This is a docstring for a python function.
+
+                   This is the summary.
+
+                   Parameters
+                   ----------
+
+                   param1
+                       Description for param1.
+
+                   param2
+                       Description for param2.
+
+                   Returns
+                   -------
+
+                       Absolutely nothing.
+
+                   \"\"\"
+                   pass
+           """
+
+    file = textwrap.dedent(file).lstrip()
+    fixed = textwrap.dedent(fixed).lstrip()
+    args = ["-t", "py", "-r", file]
     result = runner.invoke(main, args=args)
     assert result.exit_code == 0
     assert result.output == fixed